+++ model_bakery/recipe.py
@@ -34,19 +34,15 @@ class Recipe(Generic[M]):
         # _iterator_backups will hold values of the form (backup_iterator, usable_iterator).
         self._iterator_backups = {}  # type: Dict[str, Any]
 
-    def _mapping(  # noqa: C901
-        self, _using: str, new_attrs: Dict[str, Any]
-    ) -> Dict[str, Any]:
+    def _mapping(self, _using: str, new_attrs: Dict[str, Any]) -> Dict[str, Any]:
         _save_related = new_attrs.get("_save_related", True)
-        _quantity = new_attrs.get("_quantity")
-        if _quantity is None:
-            _quantity = 1
+        _quantity = new_attrs.get("_quantity", 1)
         rel_fields_attrs = {k: v for k, v in new_attrs.items() if "__" in k}
         new_attrs = {k: v for k, v in new_attrs.items() if "__" not in k}
         mapping = self.attr_mapping.copy()
         for k, v in self.attr_mapping.items():
             # do not generate values if field value is provided
-            if new_attrs.get(k):
+            if k in new_attrs:
                 continue
             elif baker.is_iterator(v):
                 if isinstance(self._model, str):
@@ -59,24 +55,20 @@ class Recipe(Generic[M]):
                     )
                 mapping[k] = self._iterator_backups[k][1]
             elif isinstance(v, RecipeForeignKey):
-                a = {}
-                for key, _value in list(rel_fields_attrs.items()):
-                    if key.startswith(f"{k}__"):
-                        a[key] = rel_fields_attrs.pop(key)
-                recipe_attrs = baker.filter_rel_attrs(k, **a)
+                rel_attrs = baker.filter_rel_attrs(k, **rel_fields_attrs)
                 if _save_related:
                     # Create a unique foreign key for each quantity if one_to_one required
                     if v.one_to_one is True:
                         rel_gen = [
-                            v.recipe.make(_using=_using, **recipe_attrs)
+                            v.recipe.make(_using=_using, **rel_attrs)
                             for _ in range(_quantity)
                         ]
                         mapping[k] = itertools.cycle(rel_gen)
                     # Otherwise create shared foreign key for each quantity
                     else:
-                        mapping[k] = v.recipe.make(_using=_using, **recipe_attrs)
+                        mapping[k] = v.recipe.make(_using=_using, **rel_attrs)
                 else:
-                    mapping[k] = v.recipe.prepare(_using=_using, **recipe_attrs)
+                    mapping[k] = v.recipe.prepare(_using=_using, **rel_attrs)
             elif isinstance(v, related):
                 mapping[k] = v.make()
         mapping.update(new_attrs)